Switch to node-config and sub-commands
This is a big re-write and switches to using node-config for managing
the configuration. This makes it easier to have nested config and adds
support for multiple environments.

Server configuration (which will often be shared across projects) is
stored in the user's home directory.

This also adds a guided setup which prompts for the values and checks
they are good.