Allow external tool details to be looked up.
This allows the LTI tool JSON for a registration ID to be looked up.

It also checks that a tool is installed in the sub-account when updating the configuration and if it isn't adds it.

The validate command also now checks to see that the developer key is added to the sub-account.