Allow value to be empty.
For some values we want to allow empty values and only reject undefined values. An example of this is the `title_suffix` where for all instances apart from production we append something onto the name of the tool to make it easier to identify, but in production we append an empty string.