Add support for $CURRENT_DATETIME and notes
This provides an automatically expanded variable `$CURRENT_DATETIME` which is set to the current timestamp. This is useful for including a changing value in the registrations so it's easy to tell if Canvas has been updated recently.

This also supports the notes field both API and LTI keys in Canvas which is a good place to include a timestamp.